Use prebuilt responses if available
Prebuilt example responses are now served if they exist for the requested path. If several exist they are graded and the server attempts to pick the best match to send.

We don't try to filter out error responses here but it's a given that some of them will never get encountered (these are specifically the ones that don't match full path routing).

It is assumed that there is only one path with one method in each example file (because this is true of all the examples we have).