Don't allow a contract to terminate the matching flow when it can't match
The issue is that in a scenario where multiple contracts have a conflicting basePath the first one to fail will send a 404 even when a subsequent one would match. So if a request came in for "/api/foo/bar/" and contract A had basePath="/api" and contract B had basePath="/api/foo" and "/bar/" method existed on B: The request would fail because contract A would try to match it, fail, and send a 404.